Tesla: adapt to settings package move
