+++ vehicle/tesla/service.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 
 	"github.com/evcc-io/evcc/core/keys"
-	"github.com/evcc-io/evcc/server/db/settings"
+	"github.com/evcc-io/evcc/db/settings"
 	"github.com/evcc-io/evcc/server/network"
 	"github.com/evcc-io/evcc/server/service"
 )
